requirements.txt
fastapi
websockets==10.4
python-multipart
python-testui
python-dotenv
aiohttp
git+https://github.com/BarkoAgent/ba-sdk.git
